feat(username): Usernames are now mandatory.
Users are now required to choose a username before being allowed to access the rest of the page and they can now be implemented successfully. Also a lot of major refactoring when it comes to making code more readable and reusable. Specifically in the signin/signout ones and all the forms.